Refactored OER and UPER C generator: Moved C functions to separate files to make generator code clearer
